+++ process_mining/src/core/event_data/object_centric/ocel_sql/sqlite/sqlite_ocel_export.rs
@@ -13,6 +13,9 @@ pub fn export_ocel_sqlite_to_path<P: AsRef<std::path::Path>>(
     ocel: &OCEL,
     path: P,
 ) -> Result<(), DatabaseError> {
+    if path.as_ref().exists() {
+        let _ = std::fs::remove_file(&path);
+    }
     let con = Connection::open(path)?;
     export_ocel_to_sql_con(&con, ocel)
 }